LayoutTests/storage/indexeddb/modern/deletedatabase-request-event-expected.txt

ALERT: [object IDBOpenDBRequest]
ALERT: success
ALERT: Done
This test calls deleteDatabase on window.indexedDB and verifies that an event is fired at the request.
